+++ src/plot_page/__init__.py
@@ -3,7 +3,7 @@ from dash import Input, Output, State
 
 from plot_page.app import app
 from plot_page.control.data_operations import prepare_upload_data
-from plot_page.view.layout import page_layout  # noqa: F811
+from plot_page.view.components import page_layout  # noqa: F811
 from plot_page.view.pages import home, plot_2d
 
 if __name__ == "__main__":
@@ -11,6 +11,7 @@ if __name__ == "__main__":
     app.run_server(debug=True)
 
 
+#####################################################################################################################################################
 @app.callback(
     Output("page-content", "children"),
     Input("url", "pathname"),
