+++ src/plot_page/control/generate_data.py
@@ -1,9 +1,12 @@
+"""File to generate an example dataset."""
+
 import os
 import random
 
 from plot_page.control.base_functions import write_json
 
 
+#####################################################################################################################################################
 def generate_data(id: int, **kwargs) -> dict:
     """Generate a random dict.
 
@@ -15,7 +18,7 @@ def generate_data(id: int, **kwargs) -> dict:
     """
     result = {"id": id}
     for key, value in kwargs.items():
-        if  value.get("type", None) == "float":
+        if value.get("type", None) == "float":
             result[key] = random.uniform(value.get("start", 0), value.get("end", 1))
         if value.get("type", None) == "int":
             result[key] = random.randint(value.get("start", 0), value.get("end", 1))
@@ -23,7 +26,9 @@ def generate_data(id: int, **kwargs) -> dict:
             result[key] = random.choice(value["options"])
     return result
 
-def generate_testdata( number_elements: int ,**kwargs) -> list[dict]:
+
+#####################################################################################################################################################
+def generate_testdata(number_elements: int, **kwargs) -> list[dict]:
     """Function to generate a list of random data.
 
     Args:
@@ -32,8 +37,16 @@ def generate_testdata( number_elements: int ,**kwargs) -> list[dict]:
     Returns:
         list[dict]: The created random dataset.
     """
-    return [generate_data( i, **kwargs) for i in range(number_elements)]
+    return [generate_data(i, **kwargs) for i in range(number_elements)]
 
 
-test = generate_testdata(4000, type={"options": ["BMW", "VW", "Audi"]}, iteration={"type":"int", "start": 0, "end": 30}, ps={"type": "int", "start": 50, "end": 150}, km={"type": "float", "start": 0, "end": 200000}, backlights= {"type": "int"}, wheels={"type": "int"})
+test = generate_testdata(
+    4000,
+    type={"options": ["BMW", "VW", "Audi"]},
+    iteration={"type": "int", "start": 0, "end": 30},
+    ps={"type": "int", "start": 50, "end": 150},
+    km={"type": "float", "start": 0, "end": 200000},
+    backlights={"type": "int"},
+    wheels={"type": "int"},
+)
 write_json(os.path.join(".", "test2.json"), {"input2": test})
