+++ src/plot_page/control/plot_functions.py
@@ -1,3 +1,5 @@
+"""Functions that supports plotting."""
+
 import pandas as pd
 import plotly.graph_objects as go
 from dash import dcc
