+++ src/plot_page/view/pages/__init__.py
@@ -1,0 +1,1 @@
+"""Module that contains all pages."""
